[cgo][test-libc] Refs #2 Repair error in GoInt in SKY_coin_CreateUnspents and SKY_coin_CreateUnspent

+++ include/skycriterion.h
@@ -11,7 +11,7 @@ extern int isPubKeyEq(cipher__PubKey *pubkey1, cipher__PubKey *pubkey2);
 extern int isSigEq(cipher__Sig *sig1, cipher__Sig *sig2);
 
 extern int isSecKeyEq(cipher__SecKey *seckey1, cipher__SecKey *seckey2);
-extern int isU8Eq(unsigned char p1[], unsigned char p2[], int len);
+extern int isU8Eq(unsigned char p1[], unsigned char p2[], size_t len);
 
 extern int isGoSliceEq(GoSlice *slice1, GoSlice *slice2);
 extern int isGoSlice_Eq(GoSlice_ *slice1, GoSlice_ *slice2);
